// src/constant/enums.ts
export enum Role {
  ADMIN = "admin",
  USER = "user",
}

export enum Pages {
  DASHBOARD = "dashboard",
  FACILITIES = "facilities",
  ADS = "ads",
  ROOM = "rooms",
  ROOMDATA = "roomData",
  BOOKING = "booking",
  USERS = "users",
}

export enum Auth {
  LOGIN = "login",
  REGISTER = "register",
  FORGETPASS = "forgetpass",
  RESETPASS = "resetpass",
}

export enum Paths {
  LANDING = "/",
  EXPLORE = "explore",
  DETAILS = "details",
  PAYMENT = "payment",
  FAVORITE = "favorites",
}